requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
firebase-admin==6.2.0
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.2
gunicorn==21.2.0
email-validator==2.0.0 